requirements.txt
Flask
Tensorflow
Cassandra-driver
Image
Cluster
